Fix imports, refactor builders, invert feature flag
