+++ maplibre/src/context.rs
@@ -2,12 +2,14 @@ use std::ops::Div;
 
 use cgmath::Angle;
 
+use crate::render::Renderer;
+use crate::style::Style;
+use crate::window::WindowSize;
 use crate::{
     coords::{LatLon, ViewRegion, WorldCoords, Zoom, ZoomLevel, TILE_SIZE},
     io::tile_repository::TileRepository,
     render::camera::{Camera, Perspective, ViewProjection},
     util::ChangeObserver,
-    Renderer, Style, WindowSize,
 };
 
 /// Stores the camera configuration.
