+++ maplibre/src/io/scheduler.rs
@@ -7,7 +7,7 @@ use crate::error::Error;
 /// Async/await scheduler.
 /// Can schedule a task from a future factory and a shared state.
 pub trait Scheduler: 'static {
-    #[cfg(not(feature = "no-thread-safe-futures"))]
+    #[cfg(feature = "thread-safe-futures")]
     fn schedule<T>(
         &self,
         future_factory: impl (FnOnce() -> T) + Send + 'static,
@@ -15,7 +15,7 @@ pub trait Scheduler: 'static {
     where
         T: Future<Output = ()> + Send + 'static;
 
-    #[cfg(feature = "no-thread-safe-futures")]
+    #[cfg(not(feature = "thread-safe-futures"))]
     fn schedule<T>(
         &self,
         future_factory: impl (FnOnce() -> T) + Send + 'static,
