+++ maplibre/src/platform/noweb/scheduler.rs
@@ -1,6 +1,7 @@
 use std::future::Future;
 
-use crate::{error::Error, Scheduler};
+use crate::error::Error;
+use crate::io::scheduler::Scheduler;
 
 /// Multi-threading with Tokio.
 pub struct TokioScheduler;
