+++ maplibre/src/render/resource/surface.rs
@@ -3,10 +3,10 @@
 
 use std::{mem::size_of, sync::Arc};
 
+use crate::window::{MapWindow, WindowSize};
 use crate::{
     render::{eventually::HasChanged, resource::texture::TextureView, settings::RendererSettings},
     window::HeadedMapWindow,
-    MapWindow, WindowSize,
 };
 
 pub struct BufferDimensions {
