+++ maplibre/src/render/stages/graph_runner_stage.rs
@@ -2,13 +2,13 @@
 
 use log::error;
 
+use crate::render::Renderer;
 use crate::{
     context::MapContext,
     render::{
         eventually::Eventually::Initialized, graph::RenderGraph, graph_runner::RenderGraphRunner,
     },
     schedule::Stage,
-    Renderer,
 };
 
 /// Updates the [`RenderGraph`] with all of its nodes and then runs it to render the entire frame.
