+++ maplibre/src/render/stages/phase_sort_stage.rs
@@ -1,6 +1,7 @@
 //! Sorts items of the [RenderPhases](RenderPhase).
 
-use crate::{context::MapContext, render::render_phase::RenderPhase, schedule::Stage, Renderer};
+use crate::render::Renderer;
+use crate::{context::MapContext, render::render_phase::RenderPhase, schedule::Stage};
 
 #[derive(Default)]
 pub struct PhaseSortStage;
