+++ maplibre/src/render/stages/queue_stage.rs
@@ -1,12 +1,12 @@
 //! Queues [PhaseItems](crate::render::render_phase::PhaseItem) for rendering.
 
+use crate::render::{RenderState, Renderer};
 use crate::{
     context::MapContext,
     render::{
         eventually::Eventually::Initialized, resource::IndexEntry, tile_view_pattern::TileInView,
     },
     schedule::Stage,
-    RenderState, Renderer,
 };
 
 #[derive(Default)]
