+++ web/Cargo.toml
@@ -19,7 +19,7 @@ crate-type = ["cdylib", "rlib"]
 
 [dependencies]
 async-trait = "0.1.56"
-maplibre = { path = "../maplibre", features = ["no-thread-safe-futures"] }
+maplibre = { path = "../maplibre", default-features = false, features = [] }
 maplibre-winit = { path = "../maplibre-winit", version = "0.0.1"  }
 
 log = "0.4.17"
